+++ src/idc/imgaug/filter/__init__.py
@@ -4,6 +4,7 @@ from ._base_image_augmentation import BaseImageAugmentation
 from ._change_grayscale import ChangeGrayscale
 from ._clip_grayscale import ClipGrayscale
 from ._crop import Crop
+from ._crop_to_label import CropToLabel
 from ._enhance import Enhance, ENHANCEMENTS, ENHANCEMENT_CONTRAST, ENHANCEMENT_COLOR, ENHANCEMENT_SHARPNESS, ENHANCEMENT_BRIGHTNESS
 from ._find_contours import FindContours, CONNECTIVITY, CONNECTIVITY_LOW, CONNECTIVITY_HIGH
 from ._flip import Flip